feat(protocol): host ApiErrorCode + client tRPC error helpers under api-schemas subpath
