+++ packages/protocol/src/api-schemas/api-error-code.ts
@@ -1,0 +1,24 @@
+export enum ApiErrorCode {
+  NOT_FOUND = "NOT_FOUND",
+  ALREADY_EXISTS = "ALREADY_EXISTS",
+  INVALID_INPUT = "INVALID_INPUT",
+  MISSING_REQUIRED_FIELD = "MISSING_REQUIRED_FIELD",
+  UNAUTHORIZED = "UNAUTHORIZED",
+  FORBIDDEN = "FORBIDDEN",
+  INTERNAL_ERROR = "INTERNAL_ERROR",
+  SERVICE_UNAVAILABLE = "SERVICE_UNAVAILABLE",
+  WORKFLOW_NOT_FOUND = "WORKFLOW_NOT_FOUND",
+  WORKFLOW_EXECUTION_FAILED = "WORKFLOW_EXECUTION_FAILED",
+  ASSET_NOT_FOUND = "ASSET_NOT_FOUND",
+  ASSET_UPLOAD_FAILED = "ASSET_UPLOAD_FAILED",
+  PYTHON_BRIDGE_UNAVAILABLE = "PYTHON_BRIDGE_UNAVAILABLE"
+}
+
+export interface ApiErrorResponse {
+  code: ApiErrorCode;
+  detail: string;
+}
+
+export function apiError(code: ApiErrorCode, detail: string): ApiErrorResponse {
+  return { code, detail };
+}
