+++ packages/protocol/src/api-schemas/index.ts
@@ -1,4 +1,2 @@
-// Zod input/output schemas for tRPC procedures.
-// One file per domain; domain-specific schema modules are re-exported as
-// namespaces below as they're added (e.g. `export * as costs from "./costs.js";`).
-export {};
+export * from "./api-error-code.js";
+export * from "./error-helpers.js";
